Close #137 - [`extras-scala-io`] Change `file.ops` to `file.syntax`
