+++ modules/extras-scala-io/shared/src/main/scala-2.11_2.12/extras/scala/io/file/TempFiles.scala
@@ -26,7 +26,7 @@ object TempFiles {
   final case class TempDir(value: File) extends AnyVal
   object TempDir {
     implicit val deleteTempFileWhenRelease: CanClose[TempDir] = new CanClose[TempDir] {
-      def close(tempDir: TempDir): Unit = ops.deleteAllRecursively(tempDir.value)
+      def close(tempDir: TempDir): Unit = syntax.deleteAllRecursively(tempDir.value)
     }
   }
 }
