+++ modules/extras-scala-io/shared/src/main/scala-2.13/extras/scala/io/file/TempFiles.scala
@@ -18,6 +18,6 @@ object TempFiles {
 
   final case class TempDir(value: File) extends AnyVal
   object TempDir {
-    implicit val deleteTempFileWhenRelease: CanClose[TempDir] = tempDir => ops.deleteAllRecursively(tempDir.value)
+    implicit val deleteTempFileWhenRelease: CanClose[TempDir] = tempDir => syntax.deleteAllRecursively(tempDir.value)
   }
 }
