+++ modules/extras-scala-io/shared/src/main/scala-3/extras/scala/io/file/TempFiles.scala
@@ -27,6 +27,6 @@ object TempFiles {
 
     given tempDirCanEqual: CanEqual[TempDir, TempDir] = CanEqual.derived
 
-    given deleteTempFileWhenRelease: CanClose[TempDir] = tempDir => ops.deleteAllRecursively(tempDir.value)
+    given deleteTempFileWhenRelease: CanClose[TempDir] = tempDir => syntax.deleteAllRecursively(tempDir.value)
   }
 }
